Revert PHF optimization and create baseline for string allocation optimization
- Removed all Perfect Hash Function (PHF) related code and dependencies
- Reverted all 14 script converters to simple HashMap implementations
- Removed generic_phf.rs, phf_optimization.rs, and related files
- Removed PHF dependencies from Cargo.toml
- Documented PHF learnings in OPTIMIZATION_ROADMAP.md
- All tests passing (197 tests)

PHF optimization showed 16.7-27.5% performance regression instead of
expected improvements. Modern HashMap<char, char> is already highly
optimized for small maps (77 characters). The complexity of PHF was
not justified by the minimal performance gains.

Ready for next optimization: string allocation reduction.

+++ examples/investigate_performance_regression.rs
@@ -86,7 +86,7 @@ fn measure_latency(shlesha: &Shlesha, from: &str, to: &str, text: &str, iteratio
         times.push(duration.as_nanos() as f64);
     }
     
-    times.iter().sum::<f64>() / times.len()
+    times.iter().sum::<f64>() / times.len() as f64
 }
 
 fn test_direct_mapping_usage(shlesha: &Shlesha) {
